masked experiments for bfloat16

+++ src/weathergen/model/attention.py
@@ -71,8 +71,8 @@ class MultiSelfAttentionHead_Varlen(torch.nn.Module):
 
         # project onto heads and q,k,v and ensure these are 4D tensors as required for flash attention
         s = [x.shape[0], self.num_heads, -1]
-        qs = self.lnorm_q(self.proj_heads_q(x).reshape(s)).to(torch.float16)
-        ks = self.lnorm_k(self.proj_heads_k(x).reshape(s)).to(torch.float16)
+        qs = self.lnorm_q(self.proj_heads_q(x).reshape(s)).to(torch.bfloat16)
+        ks = self.lnorm_k(self.proj_heads_k(x).reshape(s)).to(torch.bfloat16)
         vs = self.proj_heads_v(x).reshape(s)
 
         cum_x_lens = torch.cumsum(x_lens, 0, dtype=torch.int32)
@@ -152,8 +152,8 @@ class MultiSelfAttentionHead_Varlen_Flex(torch.nn.Module):
 
         # project onto heads and q,k,v and ensure these are 4D tensors as required for flash attention
         s = [x.shape[0], 1, self.num_heads, -1]
-        qs = self.lnorm_q(self.proj_heads_q(x).reshape(s)).to(torch.float16).permute([1, 2, 0, 3])
-        ks = self.lnorm_k(self.proj_heads_k(x).reshape(s)).to(torch.float16).permute([1, 2, 0, 3])
+        qs = self.lnorm_q(self.proj_heads_q(x).reshape(s)).to(torch.bfloat16).permute([1, 2, 0, 3])
+        ks = self.lnorm_k(self.proj_heads_k(x).reshape(s)).to(torch.bfloat16).permute([1, 2, 0, 3])
         vs = self.proj_heads_v(x).reshape(s).permute([1, 2, 0, 3])
 
         outs = self.compiled_flex_attention(qs, ks, vs).transpose(1, 2).squeeze()
@@ -227,8 +227,8 @@ class MultiSelfAttentionHead_Local(torch.nn.Module):
 
         # project onto heads
         s = [x.shape[0], x.shape[1], self.num_heads, -1]
-        qs = self.lnorm_q(self.proj_heads_q(x).reshape(s)).to(torch.float16).permute([0, 2, 1, 3])
-        ks = self.lnorm_k(self.proj_heads_k(x).reshape(s)).to(torch.float16).permute([0, 2, 1, 3])
+        qs = self.lnorm_q(self.proj_heads_q(x).reshape(s)).to(torch.bfloat16).permute([0, 2, 1, 3])
+        ks = self.lnorm_k(self.proj_heads_k(x).reshape(s)).to(torch.bfloat16).permute([0, 2, 1, 3])
         vs = self.proj_heads_v(x).reshape(s).permute([0, 2, 1, 3])
 
         outs = self.flex_attention(qs, ks, vs, block_mask=self.block_mask).transpose(1, 2)
@@ -304,9 +304,9 @@ class MultiCrossAttentionHead_Varlen(torch.nn.Module):
 
         # project onto heads and q,k,v and ensure these are 4D tensors as required for flash attention
         s = [x_q.shape[0], self.num_heads, self.dim_head_proj]
-        qs = self.lnorm_q(self.proj_heads_q(x_q).reshape(s)).to(torch.float16)
+        qs = self.lnorm_q(self.proj_heads_q(x_q).reshape(s)).to(torch.bfloat16)
         s = [x_kv.shape[0], self.num_heads, self.dim_head_proj]
-        ks = self.lnorm_k(self.proj_heads_k(x_kv).reshape(s)).to(torch.float16)
+        ks = self.lnorm_k(self.proj_heads_k(x_kv).reshape(s)).to(torch.bfloat16)
         vs = self.proj_heads_v(x_kv).reshape(s)
 
         if x_kv_lens is not None:
@@ -420,11 +420,11 @@ class MultiCrossAttentionHead_Varlen_SlicedQ(torch.nn.Module):
         # project onto heads and q,k,v and ensure these are 4D tensors as required for flash attention
         s = [x_q.shape[0], self.num_heads, self.dim_head_proj]
         qs = [
-            self.lnorm_q(head_proj(x_q_i).reshape(s)).to(torch.float16)
+            self.lnorm_q(head_proj(x_q_i).reshape(s)).to(torch.bfloat16)
             for head_proj, x_q_i in zip(self.proj_heads_q, x_q.transpose(1, 0), strict=False)
         ]
         s = [x_kv.shape[0], self.num_heads, self.dim_head_proj]
-        ks = self.lnorm_k(self.proj_heads_k(x_kv).reshape(s)).to(torch.float16)
+        ks = self.lnorm_k(self.proj_heads_k(x_kv).reshape(s)).to(torch.bfloat16)
         vs = self.proj_heads_v(x_kv).reshape(s)
 
         cum_x_q_lens = torch.cumsum(x_q_lens, 0, dtype=torch.int32)
@@ -511,8 +511,8 @@ class MultiSelfAttentionHead(torch.nn.Module):
 
         # project onto heads and q,k,v and ensure these are 4D tensors as required for flash attention
         s = [*([x.shape[0], 1] if len(x.shape) == 2 else x.shape[:-1]), self.num_heads, -1]
-        qs = self.lnorm_q(self.proj_heads_q(x).reshape(s)).to(torch.float16)
-        ks = self.lnorm_k(self.proj_heads_k(x).reshape(s)).to(torch.float16)
+        qs = self.lnorm_q(self.proj_heads_q(x).reshape(s)).to(torch.bfloat16)
+        ks = self.lnorm_k(self.proj_heads_k(x).reshape(s)).to(torch.bfloat16)
         vs = self.proj_heads_v(x).reshape(s)
 
         # ordering of tensors (seq, heads, embed) (which differs from torch's flash attention implt)
@@ -591,9 +591,9 @@ class MultiCrossAttentionHead(torch.nn.Module):
 
         # project onto heads and q,k,v and ensure these are 4D tensors as required for flash attention
         s = [x_q.shape[0], -1, self.num_heads, self.dim_head_proj]
-        qs = self.lnorm_q(self.proj_heads_q(x_q).reshape(s)).to(torch.float16).transpose(-3, -2)
+        qs = self.lnorm_q(self.proj_heads_q(x_q).reshape(s)).to(torch.bfloat16).transpose(-3, -2)
         s = [x_kv.shape[0], -1, self.num_heads, self.dim_head_proj]
-        ks = self.lnorm_k(self.proj_heads_k(x_kv).reshape(s)).to(torch.float16).transpose(-3, -2)
+        ks = self.lnorm_k(self.proj_heads_k(x_kv).reshape(s)).to(torch.bfloat16).transpose(-3, -2)
         vs = self.proj_heads_v(x_kv).reshape(s).transpose(-3, -2)
 
         # correct ordering of tensors with seq dimension second but last is critical
